Le ore sono state convertite in int

+++ model/lavoratore.cpp
@@ -1,7 +1,7 @@
 #include "lavoratore.h"
 
 Lavoratore::Lavoratore(std::string nome, std::string cognome, QDate dataNascita, std::string codiceFiscale, Genere genere, Telefono numeroTelefono, std::string reparto,
-                       float pagaPerOra, float orePreviste, float oreFerie, float orePermesso, QUuid IDAziendale):
+                       float pagaPerOra, int orePreviste, int oreFerie, int orePermesso, QUuid IDAziendale):
     Persona(nome, cognome, dataNascita, codiceFiscale, genere, numeroTelefono), _reparto(reparto), _pagaPerOra(pagaPerOra), _orePreviste(orePreviste), _oreFerie(oreFerie),
     _orePermesso(orePermesso), _IDAziendale(IDAziendale)
 {
@@ -13,17 +13,17 @@ std::string Lavoratore::getReparto() const
     return _reparto;
 }
 
-float Lavoratore::getOrePreviste() const
+int Lavoratore::getOrePreviste() const
 {
     return _orePreviste;
 }
 
-float Lavoratore::getOreFerie() const
+int Lavoratore::getOreFerie() const
 {
     return _oreFerie;
 }
 
-float Lavoratore::getOrePermesso() const
+int Lavoratore::getOrePermesso() const
 {
     return _orePermesso;
 }
