+++ model/operaio.h
@@ -10,7 +10,7 @@ class Operaio : public Lavoratore
 public:
 
     Operaio(string nome, string cognome, QDate dataNascita, string codiceFiscale, Genere genere, Telefono numeroTelefono, string reparto,
-               float pagaPerOra, float orePreviste, float oreFerie, float orePermesso, Livello livello, QUuid IDAziendale = QUuid::createUuid());
+               float pagaPerOra, int orePreviste, int oreFerie, int orePermesso, Livello livello, QUuid IDAziendale = QUuid::createUuid());
 
     Livello getLivello() const;
